Fix: Combat Camera in quick sands rooms
Combat Camera was taking "position.y" value that is the quick sand bottom, instead of "TopHeight" (formely maxceiling) that is the top surface of the quicksands camera.

When Lara digs enough in the quicksands, she automatically draws her weapons stopping the combat camera, that's why the bug only happened on certain range.